CryptoManager.Server/CryptoManager.Business/OrderBusiness.cs
using CryptoManager.Domain.Contracts.Business;
using CryptoManager.Domain.Contracts.Repositories;
using CryptoManager.Domain.Entities;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace CryptoManager.Business
{
    public class OrderBusiness : IOrderBusiness
    {
        private readonly IOrderRepository _repository;

        public OrderBusiness(IOrderRepository repository)
        {
            _repository = repository;
        }

        public Task<Order> CreateOrder(Order order)
        {
            if(order.OrderItems == null || !order.OrderItems.Any())
            {
                throw new InvalidOperationException("Order Item Must be Informed");
            }
            return _repository.InsertAsync(order);
        }
    }
}
